Allow to skip correction
no_correction passed to Corrector creates an object that returns a copy of the array when correcting.
This can be useful with extremely long input vectors when computing the correction is very expensive and the gain minimal (at least hundreds of thousands of samples and the "right" number of bins).
Use at you own risk.